+++ query/errors.go
@@ -15,5 +15,6 @@ var (
 	)
 	ErrMoreThanOneResultSet   = errors.New("unexpected more than one result set")
 	ErrNoResultSets           = errors.New("no result sets")
+	ErrOptionNotForTxExecute  = errors.New("option is not for execute on transaction")
 	ErrTxControlWithoutCommit = errors.New("read-write transaction control with BeginTx requires CommitTx to be set, otherwise data changes will not be committed to the database") //nolint:lll
 )